+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2024-01-23
+
 ### Added
 
 - Bigtable policies [#39](https://github.com/losisin/gcp-pac/pull/39)